refactor: rename step from 'fromMapToPlainObject' to 'transformMap'
